Upgrade to TestNG 7 (#5787)
* Upgrade to TestNG 7.

This fixes 'Inherited test methods do not inherit groups', https://github.com/cbeust/testng/issues/182.

* Remove unused test data class that causes TestNG to fail.
* Change test that causes TestNG to fail.

+++ src/test/java/org/broadinstitute/hellbender/engine/ProgressMeterUnitTest.java
@@ -136,7 +136,7 @@ public class ProgressMeterUnitTest extends GATKBaseTest {
                 meter.update(new SimpleInterval("1", 1, 1));
             }
 
-            Assert.assertEquals(meter.secondsSinceLastPrint(), expectedSecondsSinceLastPrint.get(i - 1), "Wrong number of seconds reported since last print");
+            Assert.assertEquals(meter.secondsSinceLastPrint(), (double) expectedSecondsSinceLastPrint.get(i - 1), "Wrong number of seconds reported since last print");
         }
 
     }
